fix single sort agg
