+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/rel/IgniteSingleSortAggregate.java
@@ -19,7 +19,6 @@ package org.apache.ignite.internal.processors.query.calcite.rel;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 import com.google.common.collect.ImmutableList;
 import org.apache.calcite.plan.RelOptCluster;
@@ -38,6 +37,8 @@ import org.apache.calcite.util.ImmutableBitSet;
 import org.apache.calcite.util.ImmutableIntList;
 import org.apache.calcite.util.Pair;
 import org.apache.ignite.internal.processors.query.calcite.trait.TraitUtils;
+
+import static org.apache.ignite.internal.processors.query.calcite.util.Commons.maxPrefix;
 
 /**
  *
@@ -75,7 +76,7 @@ public class IgniteSingleSortAggregate extends IgniteSingleAggregateBase {
 
     /** {@inheritDoc} */
     @Override public Aggregate copy(RelTraitSet traitSet, RelNode input, ImmutableBitSet groupSet, List<ImmutableBitSet> groupSets, List<AggregateCall> aggCalls) {
-        return new IgniteSingleSortAggregate(getCluster(), traitSet.replace(collation), input, groupSet, groupSets, aggCalls);
+        return new IgniteSingleSortAggregate(getCluster(), traitSet, input, groupSet, groupSets, aggCalls);
     }
 
     /** {@inheritDoc} */
@@ -110,16 +111,15 @@ public class IgniteSingleSortAggregate extends IgniteSingleAggregateBase {
     ) {
         RelCollation inputCollation = TraitUtils.collation(inputTraits.get(0));
 
-        List<RelCollation> satisfiedCollations = TraitUtils.permute(groupSet.asList()).stream()
-            .filter(col -> inputCollation.satisfies(col))
-            .collect(Collectors.toList());
+        List<Integer> newCollation = maxPrefix(inputCollation.getKeys(), groupSet.asSet());
 
-        if (satisfiedCollations.isEmpty())
+        if (newCollation.size() < groupSet.cardinality())
             return ImmutableList.of();
 
-        return satisfiedCollations.stream()
-            .map(col -> Pair.of(nodeTraits.replace(col), inputTraits))
-            .collect(Collectors.toList());
+        return ImmutableList.of(Pair.of(
+            nodeTraits.replace(RelCollations.of(ImmutableIntList.copyOf(newCollation))),
+            inputTraits
+        ));
     }
 
     /** {@inheritDoc} */
